// backend/src/statistics/dto/stock-stats.dto.ts
export class StockStatsDto {
  totalProducts: number;
  totalStock: number;
  lowStockItems: {
    productId: number;
    name: string;
    currentStock: number;
  }[];
  outOfStockItems: {
    productId: number;
    name: string;
  }[];
  byCategory: {
    categoryId: number;
    name: string;
    productCount: number;
    totalStock: number;
  }[];
}